Gerenciamento de eventos
A funcionalidade eventos funcionando completamente. CRUD e a visualização.

+++ Controller/CategoryController.js
@@ -2,6 +2,7 @@ const express = require('express');
 const slugify = require('slugify');
 const router = express.Router();
 const Category = require('../Model/Category');
+const Notice = require('../Model/Notice');
 
 //Rota de criação
 router.get('/admin/categoria' , (req,res) => {
@@ -24,14 +25,19 @@ router.post('/categoria/criar' , (req,res) => {
 
 router.get('/admin/categoria/editar/:id' , (req,res) => {
     let id = req.params.id;
-    Category.findByPk(id).then(category => {
-        res.render('category/edit',{category});
-    });
+    if(id != undefined){
+        if(!isNaN(id)){
+            Category.findByPk(id).then(category => {
+                res.render('category/edit',{category});
+            }).catch(error => {
+                res.redirect('/admin/categoria');
+            })
+        }
+    }
 });
 
 router.post('/categoria/editar' , (req,res) => {
-    let id = req.body.id;
-    let title = req.body.title;
+    let {id, title} = req.body;
     let slug = slugify(title);
     Category.update({title: slug},{where:{id}}).then(() => {
         res.redirect('/admin/categoria');
@@ -40,9 +46,17 @@ router.post('/categoria/editar' , (req,res) => {
 
 router.post('/categoria/excluir', (req,res) => {
     let id = req.body.id;
-    Category.destroy({where: {id}}).then(() => {
-        res.redirect('/admin/categoria');
-    });
+    if(id != undefined){
+        if(!isNaN(id)){
+            Notice.destroy({where: {categoryId: id}}).then(() => {
+                Category.destroy({where: {id}}).then(() => {
+                    res.redirect('/admin/categoria');
+                }).catch(error => {
+                    res.redirect('/admin/categoria');
+                });
+            })
+        }
+    }
 });
 
 module.exports = router;
